Restrict PHP version
In the [PHP required versions table](https://devdocs.prestashop.com/1.7/basics/installation/system-requirements/#php-requirements), it's mentioned that PrestaShop 1.7.6 does not support PHP 7.3. But installation happens normally if you run PrestaShop 1.7.6 with PHP 7.3, which can cause confusion with some people.

To reduce the amount of support module developers will have to spend with issues that can be solved without their inputs, I suggest to add this constraint in `composer.json` so that developers know beforehand that this configuration is not supported.